css: Fix styling for config options
The current style for config options linking to a description
in another page belonging to the sample/application is the
same as ref link style. The fix renders the style same as the
one we have for library kconfig options and for the sample
options linking to the description in the same page as
the link occurs.

Signed-off-by: Uma Praseeda <uma.praseeda@nordicsemi.no>