theme: add support for standalone mode
The NCS theme may be used for docsets that are standalone, i.e. not part
of the NCS documentation build. For such cases, users can now enable
standalone mode by setting:

html_theme_options = {"standalone": True}

When this mode is enabled, the version switcher and the docsets bar will
not be rendered. docset/docsets options won't be required either.

Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>